More rearranging of source files. Struct definitions are now in there most related .h files, even though this causes some problems with declaration order (that were fixed).
Added a -c option, to force printing the number of each record type in a
transaction.

Fixed a bug in tcp dns parsing. We can't skip AD and NS records when parsing
TCP DNS, as that will mess up our position in the stream. Added a
parameter to DNS parse to address this issue, since we don't pass the
info in any other way.